Fixed the PropertyController.
Fixed a spelling mistake.
Fixed a bug that caused the buy house menu to crash.
Fixed a bug that cause the house menu to not allow you to buy house if
you had more than 0 on a Street.
Added a feature that allows the MainController.addReturnToArray to
handle null inputs.
Fixed a bug that caused you to be able to buy more hotels on the same
Street.  
Fixed a bug that cause the player balance on the screen to not be
changed when buying houses.
Fixed a message that told you that you could buy more house, when you
had hotels on all streets.
Fixed a bug that caused only 1 colour to appear when buying property.
Fixed a bug that caused the getRent() method to return a value
corresponding to having one more house than you actually do.